Moves customer first appointment query to customer controller.
