+++ node-backend/src/controllers/appointmentController.js
@@ -26,31 +26,6 @@ class AppointmentController {
         }
     }
 
-    static async getCustomerFirstUpcomingAppointment(req, res) {
-        try {
-            let firstUpcomingAppointment = {};
-            let serviceName = "";
-            let businessName = "";
-
-            await appointment.find({ customerId: req.query.customerId, startDateTime: { $gte: req.query.onOrAfter } })
-                .then(foundAppointments => firstUpcomingAppointment = foundAppointments[0])
-                .then(serviceName = await service.findById(firstUpcomingAppointment.serviceId).name)
-                .then(businessName = await business.findById(firstUpcomingAppointment.businessId).name)
-                .then(
-                    await res.status(200).json({
-                        _id: firstUpcomingAppointment._id,
-                        serviceName: serviceName,
-                        startDateTime: firstUpcomingAppointment.startDateTime,
-                        businessName: businessName
-                    })
-                );
-        } catch (error) {
-            res.status(500).json({
-                message: "Internal server error on AppointmentController.getCustomerFirstUpcomingAppointment(): " + error
-            })
-        }
-    }
-
     static async createAppointment(req, res) {
         try {
             const createdAppointment = await appointment.create(req.body);
