+++ node-backend/src/controllers/customerController.js
@@ -1,4 +1,7 @@
 import customer from "../models/Customer.js";
+import appointment from "../models/Appointment.js";
+import service from "../models/Service.js";
+import business from "../models/Business.js";
 
 class CustomerController {
 
@@ -21,6 +24,31 @@ class CustomerController {
             res.status(500).json({
                 message: "Internal server error on CustomerController.getCustomer(): " + error.message
             });
+        }
+    }
+
+    static async getCustomerFirstUpcomingAppointment(req, res) {
+        try {
+            let firstUpcomingAppointment = {};
+            let serviceName = "";
+            let businessName = "";
+
+            await appointment.find({ customerId: req.query.customerId, startDateTime: { $gte: req.query.onOrAfter } })
+                .then(foundAppointments => firstUpcomingAppointment = foundAppointments[0])
+                .then(serviceName = await service.findById(firstUpcomingAppointment.serviceId).name)
+                .then(businessName = await business.findById(firstUpcomingAppointment.businessId).name)
+                .then(
+                    await res.status(200).json({
+                        _id: firstUpcomingAppointment._id,
+                        serviceName: serviceName,
+                        startDateTime: firstUpcomingAppointment.startDateTime,
+                        businessName: businessName
+                    })
+                );
+        } catch (error) {
+            res.status(500).json({
+                message: "Internal server error on CustomerController.getCustomerFirstUpcomingAppointment(): " + error
+            })
         }
     }
