+++ node-backend/src/routes/appointmentRoutes.js
@@ -3,7 +3,6 @@ import AppointmentController from "../controllers/appointmentController.js";
 
 const router = express.Router();
 
-router.get("/appointments", AppointmentController.getCustomerFirstUpcomingAppointment);
 router.get("/appointments", AppointmentController.getAppointments);
 router.get("/appointments/:id", AppointmentController.getAppointment);
 router.post("/appointments", AppointmentController.createAppointment);
