+++ node-backend/src/routes/customerRoutes.js
@@ -4,6 +4,7 @@ import CustomerController from "../controllers/customerController.js";
 const router = express.Router();
 
 router.get("/customers", CustomerController.getCustomers);
+router.get("/customers/:id/appointments/first", AppointmentController.getCustomerFirstUpcomingAppointment);
 router.get("/customers/:id", CustomerController.getCustomer);
 router.post("/customers", CustomerController.createCustomer);
 router.put("/customers/:id", CustomerController.updateCustomer);
